Add "From" header, set to a dummy address (nouser@matrix.net)
This header is contained in the $headers associative array, which is then passed as the 4th argument of the mail function.

This is done in order to customize the "From" E-Mail field, replacing the default "From" address set in php.ini.

However, the original sender address (which is the default E-Mail account of the Mail Server) is still present in headers such as "Envelope-from" and "Return-Path", but we don't care about changing them too.